+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Ensure default calendar is set correctly when reading XER and PMXML files, and P6 databases.
 * Use default hours per day/week/month/year when reading P6 XER files or databases if these values are not present.
 * Ensure that the minutes per day/week/month/year attributes are copied when a calendar is copied.
+* When reading P6 schedules, roll up calendar for WBS entries when child activities all share the same calendar.
 
 ## 9.6.0 (13/09/21)
 * Add Planned Start and Scheduled Finish to project properties.